+++ src/store/name-generator/hardcoded-data.ts
